+++ src/Identity/Commands/DeleteAccount.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace OG\Account\Domain\Identity\Commands;
+
+use OG\Core\Domain\Command;
+
+/**
+ * Command to delete an account.
+ */
+class DeleteAccount implements Command
+{
+}